RPM builds: fix broken RPM updates for some packages
* The old package's %postun section is executed after the new
  package has been installed. Hence it must not remove files
  unless it is executed in a clean uninstall.
* This had broken RPM updates for packages like libcorefoundation.

+++ Packaging/RedHat/SPECS/libobjc2.spec
@@ -108,7 +108,7 @@ if [ -f /usr/NextSpace/include/Block.h ];then
 fi
 
 %postun devel
-if [ -f /usr/NextSpace/include/Block-libdispatch.h ];then
+if [ $1 -eq 0 -a -f /usr/NextSpace/include/Block-libdispatch.h ];then
 	mv /usr/NextSpace/include/Block-libdispatch.h /usr/NextSpace/include/Block.h
 	rm /usr/NextSpace/include/Block.h
 fi
